Move common Exceptions to common.exceptions module
