+++ concent_api/core/message_handlers.py
@@ -20,7 +20,14 @@ from golem_messages import message
 from golem_messages.message import FileTransferToken
 from golem_messages.message.tasks import SubtaskResultsRejected
 
-from core.exceptions import ConcentInSoftShutdownMode
+from common.constants import ErrorCode
+from common.exceptions import ConcentInSoftShutdownMode
+from common.helpers import deserialize_message
+from common.helpers import get_current_utc_timestamp
+from common.helpers import parse_timestamp_to_utc_datetime
+from common.helpers import sign_message
+from common import logging
+
 from core.exceptions import Http400
 from core.models import Client
 from core.models import PaymentInfo
@@ -42,12 +49,7 @@ from core.validation import validate_golem_message_subtask_results_rejected
 from core.validation import validate_report_computed_task_time_window
 from core.validation import validate_secure_hash_algorithm
 from core.validation import validate_task_to_compute
-from common import logging
-from common.constants import ErrorCode
-from common.helpers import deserialize_message
-from common.helpers import get_current_utc_timestamp
-from common.helpers import parse_timestamp_to_utc_datetime
-from common.helpers import sign_message
+
 from .utils import hex_to_bytes_convert
 
 logger = getLogger(__name__)
